# Conflicts:
#	modules/SofaBoundaryCondition/ConicalForceField.h
#	modules/SofaBoundaryCondition/EllipsoidForceField.h
#	modules/SofaBoundaryCondition/FixedTranslationConstraint.h
#	modules/SofaBoundaryCondition/LinearVelocityConstraint.h
#	modules/SofaBoundaryCondition/PartialFixedConstraint.h
#	modules/SofaBoundaryCondition/PointConstraint.inl
#	modules/SofaBoundaryCondition/ProjectDirectionConstraint.inl
#	modules/SofaBoundaryCondition/ProjectToLineConstraint.inl
#	modules/SofaBoundaryCondition/ProjectToPlaneConstraint.inl
#	modules/SofaBoundaryCondition/ProjectToPointConstraint.inl
#	modules/SofaBoundaryCondition/SphereForceField.h
#	modules/SofaBoundaryCondition/TrianglePressureForceField.inl
#	modules/SofaConstraint/BilateralInteractionConstraint.inl
#	modules/SofaConstraint/SlidingConstraint.inl
#	modules/SofaConstraint/UnilateralInteractionConstraint.inl
#	modules/SofaGeneralObjectInteraction/InteractionEllipsoidForceField.inl
#	modules/SofaSparseSolver/src/SofaSparseSolver/SparseCholeskySolver.h

+++ SofaKernel/modules/SofaBaseTopology/GridTopology.h
@@ -168,7 +168,7 @@ public:
     virtual void parse(core::objectmodel::BaseObjectDescription* arg) override ;
 
     /// Overload Method from @sa MeshTopology::getNbHexahedra
-    virtual int getNbHexahedra() override { return (d_n.getValue()[0]-1)*(d_n.getValue()[1]-1)*(d_n.getValue()[2]-1); }
+    virtual size_t getNbHexahedra() override { return (d_n.getValue()[0]-1)*(d_n.getValue()[1]-1)*(d_n.getValue()[2]-1); }
     /// Overload Method from @sa MeshTopology::getQuad
     Quad getQuad(int x, int y, int z);
